\name{c.rgb2hex}
\alias{c.rgb2hex}
\title{Convert a rgb color represented by values between 0 and 1
in a rgb represented by Hexadecimal values}
\usage{
  c.rgb2hex(r, g, b)
}
\arguments{
  \item{r}{red (from 0 to 1)}

  \item{g}{green (from 0 to 1)}

  \item{b}{blue (from 0 to 1)}
}
\description{
  Convert a rgb color represented by values between 0 and 1
  in a rgb represented by Hexadecimal values
}

